requirements.txt
flask
pandas
numpy
openai
scikit-learn
python-dotenv
requests
matplotlib
seaborn


# --- File Loaders ---
pypdf==4.3.1

# --- Environment Variables ---
python-dotenv==1.0.1
google-generativeai==0.7.2

groq==0.11.0

# --- Utilities ---
tqdm==4.66.5
numpy==1.26.4
langchain==0.3.3
langchain-core==0.3.3
langchain-community==0.3.2
langchain-text-splitters==0.3.0
langchain-google-genai==1.0.3
langchain-groq
faiss-cpu>=1.8.0
pypdf
python-dotenv
langchain==0.3.3
langchain-core==0.3.3
langchain-community==0.3.2
langchain-text-splitters==0.3.0
langchain-google-genai==1.0.3
langchain-groq==0.1.5
faiss-cpu
pypdf
tiktoken
reportlab